Update readme.md
Hi, I'm Harini Senthilkumar, a passionate Full Stack Developer with a love for creating responsive and dynamic web applications. I enjoy working with a variety of technologies, especially Python and the MERN stack. Explore my GitHub for exciting projects, and feel free to connect with me to collaborate or ask about web development !!!